% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_rate.R
\name{get_raw_grad_rate}
\alias{get_raw_grad_rate}
\title{Get Raw Graduation Rate}
\usage{
get_raw_grad_rate(end_year)
}
\arguments{
\item{end_year}{four digit number represents the end of school year}
}
\value{
dataframe for a specific graduation rate
}
\description{
Get Raw Graduation Rate
}
